split fmreader
