--- src/all/fmreader/build.gradle
@@ -1,13 +1,0 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-
-ext {
-    extName = 'FMReader (multiple aggregators)'
-    pkgNameSuffix = 'all.fmreader'
-    extClass = '.FMReaderFactory'
-    extVersionCode = 28
-    libVersion = '1.2'
-    containsNsfw = true
-}
-
-apply from: "$rootDir/common.gradle"
